database: pull out database code into a new module
We're going to reuse the database controllers for the accounting plugin
