+++ .gitattributes
@@ -7,8 +7,8 @@ configure text eol=lf
 # The following files are generated and should not be shown in the
 # diffs by default on Github.
 doc/lightning*.7 linguist-generated=true
-wallet/db_*_sqlgen.c linguist-generated=true
-wallet/statements_gettextgen.po linguist-generated=true
+db_*_sqlgen.c linguist-generated=true
+statements_gettextgen.po linguist-generated=true
 *_wiregen.? linguist-generated=true
 *_printgen.? linguist-generated=true
