+++ lightningd/jsonrpc.c
@@ -17,6 +17,7 @@
 #include <ccan/asort/asort.h>
 #include <ccan/err/err.h>
 #include <ccan/io/io.h>
+#include <ccan/json_escape/json_escape.h>
 #include <ccan/json_out/json_out.h>
 #include <ccan/tal/str/str.h>
 #include <common/configdir.h>
@@ -26,6 +27,7 @@
 #include <common/memleak.h>
 #include <common/param.h>
 #include <common/timeout.h>
+#include <db/exec.h>
 #include <fcntl.h>
 #include <lightningd/jsonrpc.h>
 #include <lightningd/plugin_hook.h>
