+++ wallet/db.c
@@ -4,19 +4,18 @@
 #include <ccan/build_assert/build_assert.h>
 #include <ccan/mem/mem.h>
 #include <ccan/tal/str/str.h>
-#include <common/htlc_state.h>
 #include <common/key_derive.h>
-#include <common/onionreply.h>
 #include <common/version.h>
+#include <db/bindings.h>
+#include <db/common.h>
+#include <db/exec.h>
+#include <db/utils.h>
 #include <errno.h>
 #include <hsmd/hsmd_wiregen.h>
 #include <lightningd/channel.h>
 #include <lightningd/plugin_hook.h>
 #include <wallet/db.h>
-#include <wallet/db_common.h>
 #include <wire/wire_sync.h>
-
-#define NSEC_IN_SEC 1000000000
 
 /* Small container for things that are needed by migrations. The
  * fields are guaranteed to be initialized and can be relied upon when
@@ -872,337 +871,6 @@ static struct migration dbmigrations[] = {
     {SQL("ALTER TABLE channel_funding_inflights ADD lease_fee BIGINT DEFAULT 0"), NULL},
 };
 
-/* Leak tracking. */
-#if DEVELOPER
-static void db_assert_no_outstanding_statements(struct db *db)
-{
-	struct db_stmt *stmt;
-
-	stmt = list_top(&db->pending_statements, struct db_stmt, list);
-	if (stmt)
-		db_fatal("Unfinalized statement %s", stmt->location);
-}
-#else
-static void db_assert_no_outstanding_statements(struct db *db)
-{
-}
-#endif
-
-static void db_stmt_free(struct db_stmt *stmt)
-{
-	if (!stmt->executed)
-		fatal("Freeing an un-executed statement from %s: %s",
-		      stmt->location, stmt->query->query);
-#if DEVELOPER
-	/* If they never got a db_step, we don't track */
-	if (stmt->cols_used) {
-		for (size_t i = 0; i < stmt->query->num_colnames; i++) {
-			if (!stmt->query->colnames[i].sqlname)
-				continue;
-			if (!strset_get(stmt->cols_used,
-					stmt->query->colnames[i].sqlname)) {
-				log_broken(stmt->db->log,
-					   "Never accessed column %s in query %s",
-					   stmt->query->colnames[i].sqlname,
-					   stmt->query->query);
-			}
-		}
-		strset_clear(stmt->cols_used);
-	}
-#endif
-	if (stmt->inner_stmt)
-		stmt->db->config->stmt_free_fn(stmt);
-	assert(stmt->inner_stmt == NULL);
-}
-
-/* Matches the hash function used in devtools/sql-rewrite.py */
-static u32 hash_djb2(const char *str)
-{
-	u32 hash = 5381;
-	for (size_t i = 0; str[i]; i++)
-		hash = ((hash << 5) + hash) ^ str[i];
-	return hash;
-}
-
-struct db_stmt *db_prepare_v2_(const char *location, struct db *db,
-				     const char *query_id)
-{
-	struct db_stmt *stmt = tal(db, struct db_stmt);
-	size_t num_slots, pos;
-
-	/* Normalize query_id paths, because unit tests are compiled with this
-	 * prefix. */
-	if (strncmp(query_id, "./", 2) == 0)
-		query_id += 2;
-
-	if (!db->in_transaction)
-		db_fatal("Attempting to prepare a db_stmt outside of a "
-			 "transaction: %s", location);
-
-	/* Look up the query by its ID */
-	pos = hash_djb2(query_id) % db->config->query_table_size;
-	for (;;) {
-		if (!db->config->query_table[pos].name)
-			fatal("Could not resolve query %s", query_id);
-		if (streq(query_id, db->config->query_table[pos].name)) {
-			stmt->query = &db->config->query_table[pos];
-			break;
-		}
-		pos = (pos + 1) % db->config->query_table_size;
-	}
-
-	num_slots = stmt->query->placeholders;
-	/* Allocate the slots for placeholders/bindings, zeroed next since
-	 * that sets the type to DB_BINDING_UNINITIALIZED for later checks. */
-	stmt->bindings = tal_arr(stmt, struct db_binding, num_slots);
-	for (size_t i=0; i<num_slots; i++)
-		stmt->bindings[i].type = DB_BINDING_UNINITIALIZED;
-
-	stmt->location = location;
-	stmt->error = NULL;
-	stmt->db = db;
-	stmt->executed = false;
-	stmt->inner_stmt = NULL;
-
-	tal_add_destructor(stmt, db_stmt_free);
-
-	list_add(&db->pending_statements, &stmt->list);
-
-#if DEVELOPER
-	stmt->cols_used = NULL;
-#endif /* DEVELOPER */
-
-	return stmt;
-}
-
-#define db_prepare_v2(db,query) \
-	db_prepare_v2_(__FILE__ ":" stringify(__LINE__), db, query)
-
-bool db_step(struct db_stmt *stmt)
-{
-	bool ret;
-
-	assert(stmt->executed);
-	ret = stmt->db->config->step_fn(stmt);
-
-#if DEVELOPER
-	/* We only track cols_used if we return a result! */
-	if (ret && !stmt->cols_used) {
-		stmt->cols_used = tal(stmt, struct strset);
-		strset_init(stmt->cols_used);
-	}
-#endif
-	return ret;
-}
-
-size_t db_count_changes(struct db_stmt *stmt)
-{
-	assert(stmt->executed);
-	return stmt->db->config->count_changes_fn(stmt);
-}
-
-u64 db_last_insert_id_v2(struct db_stmt *stmt TAKES)
-{
-	u64 id;
-	assert(stmt->executed);
-	id = stmt->db->config->last_insert_id_fn(stmt);
-
-	if (taken(stmt))
-		tal_free(stmt);
-
-	return id;
-}
-
-static void destroy_db(struct db *db)
-{
-	db_assert_no_outstanding_statements(db);
-
-	if (db->config->teardown_fn)
-		db->config->teardown_fn(db);
-}
-
-/* We expect min changes (ie. BEGIN TRANSACTION): report if more.
- * Optionally add "final" at the end (ie. COMMIT). */
-static void db_report_changes(struct db *db, const char *final, size_t min)
-{
-	assert(db->changes);
-	assert(tal_count(db->changes) >= min);
-
-	/* Having changes implies that we have a dirty TX. The opposite is
-	 * currently not true, e.g., the postgres driver doesn't record
-	 * changes yet. */
-	assert(!tal_count(db->changes) || db->dirty);
-
-	if (tal_count(db->changes) > min)
-		plugin_hook_db_sync(db);
-	db->changes = tal_free(db->changes);
-}
-
-static void db_prepare_for_changes(struct db *db)
-{
-	assert(!db->changes);
-	db->changes = tal_arr(db, const char *, 0);
-}
-
-bool db_in_transaction(struct db *db)
-{
-	return db->in_transaction;
-}
-
-void db_begin_transaction_(struct db *db, const char *location)
-{
-	bool ok;
-	if (db->in_transaction)
-		db_fatal("Already in transaction from %s", db->in_transaction);
-
-	/* No writes yet. */
-	db->dirty = false;
-
-	db_prepare_for_changes(db);
-	ok = db->config->begin_tx_fn(db);
-	if (!ok)
-		db_fatal("Failed to start DB transaction: %s", db->error);
-
-	db->in_transaction = location;
-}
-
-/* By making the update conditional on the current value we expect we
- * are implementing an optimistic lock: if the update results in
- * changes on the DB we know that the data_version did not change
- * under our feet and no other transaction ran in the meantime.
- *
- * Notice that this update effectively locks the row, so that other
- * operations attempting to change this outside the transaction will
- * wait for this transaction to complete. The external change will
- * ultimately fail the changes test below, it'll just delay its abort
- * until our transaction is committed.
- */
-static void db_data_version_incr(struct db *db)
-{
-       struct db_stmt *stmt = db_prepare_v2(
-	       db, SQL("UPDATE vars "
-		       "SET intval = intval + 1 "
-		       "WHERE name = 'data_version'"
-		       " AND intval = ?"));
-       db_bind_int(stmt, 0, db->data_version);
-       db_exec_prepared_v2(stmt);
-       if (db_count_changes(stmt) != 1)
-	       fatal("Optimistic lock on the database failed. There may be a "
-                     "concurrent access to the database. Aborting since "
-                     "concurrent access is unsafe.");
-       tal_free(stmt);
-       db->data_version++;
-}
-
-void db_commit_transaction(struct db *db)
-{
-	bool ok;
-	assert(db->in_transaction);
-	db_assert_no_outstanding_statements(db);
-
-	/* Increment before reporting changes to an eventual plugin. */
-	if (db->dirty)
-		db_data_version_incr(db);
-
-	db_report_changes(db, NULL, 0);
-	ok = db->config->commit_tx_fn(db);
-
-	if (!ok)
-		db_fatal("Failed to commit DB transaction: %s", db->error);
-
-	db->in_transaction = NULL;
-	db->dirty = false;
-}
-
-static struct db_config *db_config_find(const char *dsn)
-{
-	size_t num_configs;
-	struct db_config **configs = autodata_get(db_backends, &num_configs);
-	const char *sep, *driver_name;
-	sep = strstr(dsn, "://");
-
-	if (!sep)
-		db_fatal("%s doesn't look like a valid data-source name (missing \"://\" separator.", dsn);
-
-	driver_name = tal_strndup(tmpctx, dsn, sep - dsn);
-
-	for (size_t i=0; i<num_configs; i++) {
-		if (streq(driver_name, configs[i]->name)) {
-			tal_free(driver_name);
-			return configs[i];
-		}
-	}
-
-	tal_free(driver_name);
-	return NULL;
-}
-
-/**
- * db_open - Open or create a sqlite3 database
- */
-static struct db *db_open(const tal_t *ctx, char *filename)
-{
-	struct db *db;
-
-	db = tal(ctx, struct db);
-	db->filename = tal_strdup(db, filename);
-	list_head_init(&db->pending_statements);
-	if (!strstr(db->filename, "://"))
-		db_fatal("Could not extract driver name from \"%s\"", db->filename);
-
-	db->config = db_config_find(db->filename);
-	if (!db->config)
-		db_fatal("Unable to find DB driver for %s", db->filename);
-
-	tal_add_destructor(db, destroy_db);
-	db->in_transaction = NULL;
-	db->changes = NULL;
-
-	/* This must be outside a transaction, so catch it */
-	assert(!db->in_transaction);
-
-	db_prepare_for_changes(db);
-	if (db->config->setup_fn && !db->config->setup_fn(db))
-		fatal("Error calling DB setup: %s", db->error);
-	db_report_changes(db, NULL, 0);
-
-	return db;
-}
-
-/**
- * db_get_version - Determine the current DB schema version
- *
- * Will attempt to determine the current schema version of the
- * database @db by querying the `version` table. If the table does not
- * exist it'll return schema version -1, so that migration 0 is
- * applied, which should create the `version` table.
- */
-static int db_get_version(struct db *db)
-{
-	int res = -1;
-	struct db_stmt *stmt = db_prepare_v2(db, SQL("SELECT version FROM version LIMIT 1"));
-
-	/*
-	 * Tentatively execute a query, but allow failures. Some databases
-	 * like postgres will terminate the DB transaction if there is an
-	 * error during the execution of a query, e.g., trying to access a
-	 * table that doesn't exist yet, so we need to terminate and restart
-	 * the DB transaction.
-	 */
-	if (!db_query_prepared(stmt)) {
-		db_commit_transaction(stmt->db);
-		db_begin_transaction(stmt->db);
-		tal_free(stmt);
-		return res;
-	}
-
-	if (db_step(stmt))
-		res = db_col_int(stmt, "version");
-
-	tal_free(stmt);
-	return res;
-}
-
 /**
  * db_migrate - Apply all remaining migrations from the current version
  */
@@ -1221,12 +889,12 @@ static bool db_migrate(struct lightningd *ld, struct db *db,
 	available = ARRAY_SIZE(dbmigrations) - 1;
 
 	if (current == -1)
-		log_info(db->log, "Creating database");
+		log_info(ld->log, "Creating database");
 	else if (available < current)
 		db_fatal("Refusing to migrate down from version %u to %u",
 			 current, available);
 	else if (current != available)
-		log_info(db->log, "Updating database from version %u to %u",
+		log_info(ld->log, "Updating database from version %u to %u",
 			 current, available);
 
 	while (current < available) {
@@ -1260,24 +928,13 @@ static bool db_migrate(struct lightningd *ld, struct db *db,
 	return current != orig;
 }
 
-u32 db_data_version_get(struct db *db)
-{
-	struct db_stmt *stmt;
-	u32 version;
-	stmt = db_prepare_v2(db, SQL("SELECT intval FROM vars WHERE name = 'data_version'"));
-	db_query_prepared(stmt);
-	db_step(stmt);
-	version = db_col_int(stmt, "intval");
-	tal_free(stmt);
-	return version;
-}
-
 struct db *db_setup(const tal_t *ctx, struct lightningd *ld,
 		    const struct ext_key *bip32_base)
 {
 	struct db *db = db_open(ctx, ld->wallet_dsn);
 	bool migrated;
-	db->log = new_log(db, ld->log_book, NULL, "database");
+
+	db->report_changes_fn = plugin_hook_db_sync;
 
 	db_begin_transaction(db);
 
@@ -1293,44 +950,6 @@ struct db *db_setup(const tal_t *ctx, struct lightningd *ld,
 		db_fatal("Error vacuuming db: %s", db->error);
 
 	return db;
-}
-
-s64 db_get_intvar(struct db *db, char *varname, s64 defval)
-{
-	s64 res = defval;
-	struct db_stmt *stmt = db_prepare_v2(
-	    db, SQL("SELECT intval FROM vars WHERE name= ? LIMIT 1"));
-	db_bind_text(stmt, 0, varname);
-	if (!db_query_prepared(stmt))
-		goto done;
-
-	if (db_step(stmt))
-		res = db_col_int(stmt, "intval");
-
-done:
-	tal_free(stmt);
-	return res;
-}
-
-void db_set_intvar(struct db *db, char *varname, s64 val)
-{
-	size_t changes;
-	struct db_stmt *stmt = db_prepare_v2(db, SQL("UPDATE vars SET intval=? WHERE name=?;"));
-	db_bind_int(stmt, 0, val);
-	db_bind_text(stmt, 1, varname);
-	if (!db_exec_prepared_v2(stmt))
-		db_fatal("Error executing update: %s", stmt->error);
-	changes = db_count_changes(stmt);
-	tal_free(stmt);
-
-	if (changes == 0) {
-		stmt = db_prepare_v2(db, SQL("INSERT INTO vars (name, intval) VALUES (?, ?);"));
-		db_bind_text(stmt, 0, varname);
-		db_bind_int(stmt, 1, val);
-		if (!db_exec_prepared_v2(stmt))
-			db_fatal("Error executing insert: %s", stmt->error);
-		tal_free(stmt);
-	}
 }
 
 /* Will apply the current config fee settings to all channels */
@@ -1769,618 +1388,4 @@ void migrate_last_tx_to_psbt(struct lightningd *ld, struct db *db,
 	}
 
 	tal_free(stmt);
-}
-
-void db_bind_null(struct db_stmt *stmt, int pos)
-{
-	assert(pos < tal_count(stmt->bindings));
-	stmt->bindings[pos].type = DB_BINDING_NULL;
-}
-
-void db_bind_int(struct db_stmt *stmt, int pos, int val)
-{
-	assert(pos < tal_count(stmt->bindings));
-	memcheck(&val, sizeof(val));
-	stmt->bindings[pos].type = DB_BINDING_INT;
-	stmt->bindings[pos].v.i = val;
-}
-
-void db_bind_u64(struct db_stmt *stmt, int pos, u64 val)
-{
-	memcheck(&val, sizeof(val));
-	assert(pos < tal_count(stmt->bindings));
-	stmt->bindings[pos].type = DB_BINDING_UINT64;
-	stmt->bindings[pos].v.u64 = val;
-}
-
-void db_bind_blob(struct db_stmt *stmt, int pos, const u8 *val, size_t len)
-{
-	assert(pos < tal_count(stmt->bindings));
-	stmt->bindings[pos].type = DB_BINDING_BLOB;
-	stmt->bindings[pos].v.blob = memcheck(val, len);
-	stmt->bindings[pos].len = len;
-}
-
-void db_bind_text(struct db_stmt *stmt, int pos, const char *val)
-{
-	assert(pos < tal_count(stmt->bindings));
-	stmt->bindings[pos].type = DB_BINDING_TEXT;
-	stmt->bindings[pos].v.text = val;
-	stmt->bindings[pos].len = strlen(val);
-}
-
-void db_bind_preimage(struct db_stmt *stmt, int pos, const struct preimage *p)
-{
-	db_bind_blob(stmt, pos, p->r, sizeof(struct preimage));
-}
-
-void db_bind_sha256(struct db_stmt *stmt, int pos, const struct sha256 *s)
-{
-	db_bind_blob(stmt, pos, s->u.u8, sizeof(struct sha256));
-}
-
-void db_bind_sha256d(struct db_stmt *stmt, int pos, const struct sha256_double *s)
-{
-	db_bind_sha256(stmt, pos, &s->sha);
-}
-
-void db_bind_secret(struct db_stmt *stmt, int pos, const struct secret *s)
-{
-	assert(sizeof(s->data) == 32);
-	db_bind_blob(stmt, pos, s->data, sizeof(s->data));
-}
-
-void db_bind_secret_arr(struct db_stmt *stmt, int col, const struct secret *s)
-{
-	size_t num = tal_count(s), elsize = sizeof(s->data);
-	u8 *ser = tal_arr(stmt, u8, num * elsize);
-
-	for (size_t i = 0; i < num; ++i)
-		memcpy(ser + i * elsize, &s[i], elsize);
-
-	db_bind_blob(stmt, col, ser, tal_count(ser));
-}
-
-void db_bind_txid(struct db_stmt *stmt, int pos, const struct bitcoin_txid *t)
-{
-	db_bind_sha256d(stmt, pos, &t->shad);
-}
-
-void db_bind_channel_id(struct db_stmt *stmt, int pos, const struct channel_id *id)
-{
-	db_bind_blob(stmt, pos, id->id, sizeof(id->id));
-}
-
-void db_bind_node_id(struct db_stmt *stmt, int pos, const struct node_id *id)
-{
-	db_bind_blob(stmt, pos, id->k, sizeof(id->k));
-}
-
-void db_bind_node_id_arr(struct db_stmt *stmt, int col,
-			 const struct node_id *ids)
-{
-	/* Copy into contiguous array: ARM will add padding to struct node_id! */
-	size_t n = tal_count(ids);
-	u8 *arr = tal_arr(stmt, u8, n * sizeof(ids[0].k));
-
-	for (size_t i = 0; i < n; ++i) {
-		assert(node_id_valid(&ids[i]));
-		memcpy(arr + sizeof(ids[i].k) * i,
-		       ids[i].k,
-		       sizeof(ids[i].k));
-	}
-	db_bind_blob(stmt, col, arr, tal_count(arr));
-}
-
-void db_bind_pubkey(struct db_stmt *stmt, int pos, const struct pubkey *pk)
-{
-	u8 *der = tal_arr(stmt, u8, PUBKEY_CMPR_LEN);
-	pubkey_to_der(der, pk);
-	db_bind_blob(stmt, pos, der, PUBKEY_CMPR_LEN);
-}
-
-void db_bind_short_channel_id(struct db_stmt *stmt, int col,
-			      const struct short_channel_id *id)
-{
-	char *ser = short_channel_id_to_str(stmt, id);
-	db_bind_text(stmt, col, ser);
-}
-
-void db_bind_short_channel_id_arr(struct db_stmt *stmt, int col,
-				  const struct short_channel_id *id)
-{
-	u8 *ser = tal_arr(stmt, u8, 0);
-	size_t num = tal_count(id);
-
-	for (size_t i = 0; i < num; ++i)
-		towire_short_channel_id(&ser, &id[i]);
-
-	db_bind_talarr(stmt, col, ser);
-}
-
-void db_bind_signature(struct db_stmt *stmt, int col,
-		       const secp256k1_ecdsa_signature *sig)
-{
-	u8 *buf = tal_arr(stmt, u8, 64);
-	int ret = secp256k1_ecdsa_signature_serialize_compact(secp256k1_ctx,
-							      buf, sig);
-	assert(ret == 1);
-	db_bind_blob(stmt, col, buf, 64);
-}
-
-void db_bind_timeabs(struct db_stmt *stmt, int col, struct timeabs t)
-{
-	u64 timestamp =  t.ts.tv_nsec + (((u64) t.ts.tv_sec) * ((u64) NSEC_IN_SEC));
-	db_bind_u64(stmt, col, timestamp);
-}
-
-void db_bind_tx(struct db_stmt *stmt, int col, const struct wally_tx *tx)
-{
-	u8 *ser = linearize_wtx(stmt, tx);
-	assert(ser);
-	db_bind_talarr(stmt, col, ser);
-}
-
-void db_bind_psbt(struct db_stmt *stmt, int col, const struct wally_psbt *psbt)
-{
-	size_t bytes_written;
-	const u8 *ser = psbt_get_bytes(stmt, psbt, &bytes_written);
-	assert(ser);
-	db_bind_blob(stmt, col, ser, bytes_written);
-}
-
-void db_bind_amount_msat(struct db_stmt *stmt, int pos,
-			 const struct amount_msat *msat)
-{
-	db_bind_u64(stmt, pos, msat->millisatoshis); /* Raw: low level function */
-}
-
-void db_bind_amount_sat(struct db_stmt *stmt, int pos,
-			 const struct amount_sat *sat)
-{
-	db_bind_u64(stmt, pos, sat->satoshis); /* Raw: low level function */
-}
-
-void db_bind_json_escape(struct db_stmt *stmt, int pos,
-			 const struct json_escape *esc)
-{
-	db_bind_text(stmt, pos, esc->s);
-}
-
-void db_bind_onionreply(struct db_stmt *stmt, int pos, const struct onionreply *r)
-{
-	db_bind_talarr(stmt, pos, r->contents);
-}
-
-void db_bind_talarr(struct db_stmt *stmt, int col, const u8 *arr)
-{
-	if (!arr)
-		db_bind_null(stmt, col);
-	else
-		db_bind_blob(stmt, col, arr, tal_bytelen(arr));
-}
-
-/* Local helpers once you have column number */
-static bool db_column_is_null(struct db_stmt *stmt, int col)
-{
-	return stmt->db->config->column_is_null_fn(stmt, col);
-}
-
-/* Returns true (and warns) if it's nul */
-static bool db_column_null_warn(struct db_stmt *stmt, const char *colname,
-				int col)
-{
-	if (!db_column_is_null(stmt, col))
-		return false;
-
-	log_broken(stmt->db->log, "Accessing a null column %s/%i in query %s",
-		   colname, col,
-		   stmt->query->query);
-	return true;
-}
-
-static size_t db_column_bytes(struct db_stmt *stmt, int col)
-{
-	if (db_column_is_null(stmt, col))
-		return 0;
-	return stmt->db->config->column_bytes_fn(stmt, col);
-}
-
-static const void *db_column_blob(struct db_stmt *stmt, int col)
-{
-	if (db_column_is_null(stmt, col))
-		return NULL;
-	return stmt->db->config->column_blob_fn(stmt, col);
-}
-
-
-u64 db_col_u64(struct db_stmt *stmt, const char *colname)
-{
-	size_t col = db_query_colnum(stmt, colname);
-
-	if (db_column_null_warn(stmt, colname, col))
-		return 0;
-
-	return stmt->db->config->column_u64_fn(stmt, col);
-}
-
-int db_col_int_or_default(struct db_stmt *stmt, const char *colname, int def)
-{
-	size_t col = db_query_colnum(stmt, colname);
-
-	if (db_column_is_null(stmt, col))
-		return def;
-	else
-		return stmt->db->config->column_int_fn(stmt, col);
-}
-
-int db_col_int(struct db_stmt *stmt, const char *colname)
-{
-	size_t col = db_query_colnum(stmt, colname);
-
-	if (db_column_null_warn(stmt, colname, col))
-		return 0;
-
-	return stmt->db->config->column_int_fn(stmt, col);
-}
-
-size_t db_col_bytes(struct db_stmt *stmt, const char *colname)
-{
-	size_t col = db_query_colnum(stmt, colname);
-
-	if (db_column_null_warn(stmt, colname, col))
-		return 0;
-
-	return stmt->db->config->column_bytes_fn(stmt, col);
-}
-
-int db_col_is_null(struct db_stmt *stmt, const char *colname)
-{
-	return db_column_is_null(stmt, db_query_colnum(stmt, colname));
-}
-
-const void *db_col_blob(struct db_stmt *stmt, const char *colname)
-{
-	size_t col = db_query_colnum(stmt, colname);
-
-	if (db_column_null_warn(stmt, colname, col))
-		return NULL;
-
-	return stmt->db->config->column_blob_fn(stmt, col);
-}
-
-char *db_col_strdup(const tal_t *ctx,
-		    struct db_stmt *stmt,
-		    const char *colname)
-{
-	size_t col = db_query_colnum(stmt, colname);
-
-	if (db_column_null_warn(stmt, colname, col))
-		return NULL;
-
-	return tal_strdup(ctx, (char *)stmt->db->config->column_text_fn(stmt, col));
-}
-
-void db_col_preimage(struct db_stmt *stmt, const char *colname,
-			struct preimage *preimage)
-{
-	size_t col = db_query_colnum(stmt, colname);
-	const u8 *raw;
-	size_t size = sizeof(struct preimage);
-	assert(db_column_bytes(stmt, col) == size);
-	raw = db_column_blob(stmt, col);
-	memcpy(preimage, raw, size);
-}
-
-void db_col_channel_id(struct db_stmt *stmt, const char *colname, struct channel_id *dest)
-{
-	size_t col = db_query_colnum(stmt, colname);
-
-	assert(db_column_bytes(stmt, col) == sizeof(dest->id));
-	memcpy(dest->id, db_column_blob(stmt, col), sizeof(dest->id));
-}
-
-void db_col_node_id(struct db_stmt *stmt, const char *colname, struct node_id *dest)
-{
-	size_t col = db_query_colnum(stmt, colname);
-
-	assert(db_column_bytes(stmt, col) == sizeof(dest->k));
-	memcpy(dest->k, db_column_blob(stmt, col), sizeof(dest->k));
-}
-
-struct node_id *db_col_node_id_arr(const tal_t *ctx, struct db_stmt *stmt,
-				      const char *colname)
-{
-	size_t col = db_query_colnum(stmt, colname);
-	struct node_id *ret;
-	size_t n = db_column_bytes(stmt, col) / sizeof(ret->k);
-	const u8 *arr = db_column_blob(stmt, col);
-	assert(n * sizeof(ret->k) == (size_t)db_column_bytes(stmt, col));
-	ret = tal_arr(ctx, struct node_id, n);
-
-	db_column_null_warn(stmt, colname, col);
-	for (size_t i = 0; i < n; i++)
-		memcpy(ret[i].k, arr + i * sizeof(ret[i].k), sizeof(ret[i].k));
-
-	return ret;
-}
-
-void db_col_pubkey(struct db_stmt *stmt,
-		   const char *colname,
-		   struct pubkey *dest)
-{
-	size_t col = db_query_colnum(stmt, colname);
-	bool ok;
-	assert(db_column_bytes(stmt, col) == PUBKEY_CMPR_LEN);
-	ok = pubkey_from_der(db_column_blob(stmt, col), PUBKEY_CMPR_LEN, dest);
-	assert(ok);
-}
-
-/* Yes, we put this in as a string.  Past mistakes; do not use! */
-bool db_col_short_channel_id_str(struct db_stmt *stmt, const char *colname,
-				 struct short_channel_id *dest)
-{
-	size_t col = db_query_colnum(stmt, colname);
-	const char *source = db_column_blob(stmt, col);
-	size_t sourcelen = db_column_bytes(stmt, col);
-	db_column_null_warn(stmt, colname, col);
-	return short_channel_id_from_str(source, sourcelen, dest);
-}
-
-struct short_channel_id *
-db_col_short_channel_id_arr(const tal_t *ctx, struct db_stmt *stmt, const char *colname)
-{
-	size_t col = db_query_colnum(stmt, colname);
-	const u8 *ser;
-	size_t len;
-	struct short_channel_id *ret;
-
-	db_column_null_warn(stmt, colname, col);
-	ser = db_column_blob(stmt, col);
-	len = db_column_bytes(stmt, col);
-	ret = tal_arr(ctx, struct short_channel_id, 0);
-
-	while (len != 0) {
-		struct short_channel_id scid;
-		fromwire_short_channel_id(&ser, &len, &scid);
-		tal_arr_expand(&ret, scid);
-	}
-
-	return ret;
-}
-
-bool db_col_signature(struct db_stmt *stmt, const char *colname,
-			 secp256k1_ecdsa_signature *sig)
-{
-	size_t col = db_query_colnum(stmt, colname);
-	assert(db_column_bytes(stmt, col) == 64);
-	return secp256k1_ecdsa_signature_parse_compact(
-		   secp256k1_ctx, sig, db_column_blob(stmt, col)) == 1;
-}
-
-struct timeabs db_col_timeabs(struct db_stmt *stmt, const char *colname)
-{
-	struct timeabs t;
-	u64 timestamp = db_col_u64(stmt, colname);
-	t.ts.tv_sec = timestamp / NSEC_IN_SEC;
-	t.ts.tv_nsec = timestamp % NSEC_IN_SEC;
-	return t;
-
-}
-
-struct bitcoin_tx *db_col_tx(const tal_t *ctx, struct db_stmt *stmt, const char *colname)
-{
-	size_t col = db_query_colnum(stmt, colname);
-	const u8 *src = db_column_blob(stmt, col);
-	size_t len = db_column_bytes(stmt, col);
-
-	db_column_null_warn(stmt, colname, col);
-	return pull_bitcoin_tx(ctx, &src, &len);
-}
-
-struct wally_psbt *db_col_psbt(const tal_t *ctx, struct db_stmt *stmt, const char *colname)
-{
-	size_t col = db_query_colnum(stmt, colname);
-	const u8 *src = db_column_blob(stmt, col);
-	size_t len = db_column_bytes(stmt, col);
-
-	db_column_null_warn(stmt, colname, col);
-	return psbt_from_bytes(ctx, src, len);
-}
-
-struct bitcoin_tx *db_col_psbt_to_tx(const tal_t *ctx, struct db_stmt *stmt, const char *colname)
-{
-	struct wally_psbt *psbt = db_col_psbt(ctx, stmt, colname);
-	if (!psbt)
-		return NULL;
-	return bitcoin_tx_with_psbt(ctx, psbt);
-}
-
-void *db_col_arr_(const tal_t *ctx, struct db_stmt *stmt, const char *colname,
-		  size_t bytes, const char *label, const char *caller)
-{
-	size_t col = db_query_colnum(stmt, colname);
-	size_t sourcelen;
-	void *p;
-
-	if (db_column_is_null(stmt, col))
-		return NULL;
-
-	sourcelen = db_column_bytes(stmt, col);
-
-	if (sourcelen % bytes != 0)
-		db_fatal("%s: %s/%zu column size for %zu not a multiple of %s (%zu)",
-			 caller, colname, col, sourcelen, label, bytes);
-
-	p = tal_arr_label(ctx, char, sourcelen, label);
-	memcpy(p, db_column_blob(stmt, col), sourcelen);
-	return p;
-}
-
-void db_col_amount_msat_or_default(struct db_stmt *stmt,
-				   const char *colname,
-				   struct amount_msat *msat,
-				   struct amount_msat def)
-{
-	size_t col = db_query_colnum(stmt, colname);
-
-	if (db_column_is_null(stmt, col))
-		*msat = def;
-	else
-		msat->millisatoshis = db_col_u64(stmt, colname); /* Raw: low level function */
-}
-
-void db_col_amount_msat(struct db_stmt *stmt, const char *colname,
-			struct amount_msat *msat)
-{
-	msat->millisatoshis = db_col_u64(stmt, colname); /* Raw: low level function */
-}
-
-void db_col_amount_sat(struct db_stmt *stmt, const char *colname, struct amount_sat *sat)
-{
-	sat->satoshis = db_col_u64(stmt, colname); /* Raw: low level function */
-}
-
-struct json_escape *db_col_json_escape(const tal_t *ctx,
-				       struct db_stmt *stmt, const char *colname)
-{
-	size_t col = db_query_colnum(stmt, colname);
-
-	return json_escape_string_(ctx, db_column_blob(stmt, col),
-				   db_column_bytes(stmt, col));
-}
-
-void db_col_sha256(struct db_stmt *stmt, const char *colname, struct sha256 *sha)
-{
-	size_t col = db_query_colnum(stmt, colname);
-	const u8 *raw;
-	size_t size = sizeof(struct sha256);
-	assert(db_column_bytes(stmt, col) == size);
-	raw = db_column_blob(stmt, col);
-	memcpy(sha, raw, size);
-}
-
-void db_col_sha256d(struct db_stmt *stmt, const char *colname,
-		       struct sha256_double *shad)
-{
-	size_t col = db_query_colnum(stmt, colname);
-	const u8 *raw;
-	size_t size = sizeof(struct sha256_double);
-	assert(db_column_bytes(stmt, col) == size);
-	raw = db_column_blob(stmt, col);
-	memcpy(shad, raw, size);
-}
-
-void db_col_secret(struct db_stmt *stmt, const char *colname, struct secret *s)
-{
-	size_t col = db_query_colnum(stmt, colname);
-	const u8 *raw;
-	assert(db_column_bytes(stmt, col) == sizeof(struct secret));
-	raw = db_column_blob(stmt, col);
-	memcpy(s, raw, sizeof(struct secret));
-}
-
-struct secret *db_col_secret_arr(const tal_t *ctx,
-				 struct db_stmt *stmt,
-				 const char *colname)
-{
-	return db_col_arr(ctx, stmt, colname, struct secret);
-}
-
-void db_col_txid(struct db_stmt *stmt, const char *colname, struct bitcoin_txid *t)
-{
-	db_col_sha256d(stmt, colname, &t->shad);
-}
-
-struct onionreply *db_col_onionreply(const tal_t *ctx,
-					struct db_stmt *stmt, const char *colname)
-{
-	struct onionreply *r = tal(ctx, struct onionreply);
-	r->contents = db_col_arr(ctx, stmt, colname, u8);
-	return r;
-}
-
-bool db_exec_prepared_v2(struct db_stmt *stmt TAKES)
-{
-	bool ret = stmt->db->config->exec_fn(stmt);
-
-	/* If this was a write we need to bump the data_version upon commit. */
-	stmt->db->dirty = stmt->db->dirty || !stmt->query->readonly;
-
-	stmt->executed = true;
-	list_del_from(&stmt->db->pending_statements, &stmt->list);
-
-	/* The driver itself doesn't call `fatal` since we want to override it
-	 * for testing. Instead we check here that the error message is set if
-	 * we report an error. */
-	if (!ret) {
-		assert(stmt->error);
-		db_fatal("Error executing statement: %s", stmt->error);
-	}
-
-	if (taken(stmt))
-	    tal_free(stmt);
-
-	return ret;
-}
-
-bool db_query_prepared(struct db_stmt *stmt)
-{
-	/* Make sure we don't accidentally execute a modifying query using a
-	 * read-only path. */
-	bool ret;
-	assert(stmt->query->readonly);
-	ret = stmt->db->config->query_fn(stmt);
-	stmt->executed = true;
-	list_del_from(&stmt->db->pending_statements, &stmt->list);
-	return ret;
-}
-
-void db_changes_add(struct db_stmt *stmt, const char * expanded)
-{
-	struct db *db = stmt->db;
-
-	if (stmt->query->readonly) {
-		return;
-	}
-	/* We get a "COMMIT;" after we've sent our changes. */
-	if (!db->changes) {
-		assert(streq(expanded, "COMMIT;"));
-		return;
-	}
-
-	tal_arr_expand(&db->changes, tal_strdup(db->changes, expanded));
-}
-
-const char **db_changes(struct db *db)
-{
-	return db->changes;
-}
-
-size_t db_query_colnum(const struct db_stmt *stmt,
-		       const char *colname)
-{
-	u32 col;
-
-	assert(stmt->query->colnames != NULL);
-
-	col = hash_djb2(colname) % stmt->query->num_colnames;
-	/* Will crash on NULL, which is the Right Thing */
-	while (!streq(stmt->query->colnames[col].sqlname,
-		      colname)) {
-		col = (col + 1) % stmt->query->num_colnames;
-	}
-
-#if DEVELOPER
-	strset_add(stmt->cols_used, colname);
-#endif
-
-	return stmt->query->colnames[col].val;
-}
-
-void db_col_ignore(struct db_stmt *stmt, const char *colname)
-{
-#if DEVELOPER
-	db_query_colnum(stmt, colname);
-#endif
 }
